+++ src/renderer/tr_main.c
@@ -1153,7 +1153,7 @@ static qboolean R_MirrorViewBySurface( const drawSurf_t *drawSurf, int entityNum
 		newParms.dlights = oldParms.dlights + oldParms.num_dlights;
 		newParms.num_dlights = oldParms.num_dlights;
 		r_numdlights += oldParms.num_dlights;
-		for ( i = 0; i < oldParms.num_dlights; i++ )
+		for ( i = 0; i < (int)oldParms.num_dlights; i++ )
 			newParms.dlights[i] = oldParms.dlights[i];
 	}
 #endif
@@ -1395,7 +1395,7 @@ void R_AddLitSurf( surfaceType_t *surface, shader_t *shader, int fogIndex )
 {
 	struct litSurf_s *litsurf;
 
-	if ( tr.refdef.numLitSurfs >= ARRAY_LEN( backEndData->litSurfs ) )
+	if ( (size_t)tr.refdef.numLitSurfs >= ARRAY_LEN( backEndData->litSurfs ) )
 		return;
 
 	tr.pc.c_lit_surfs++;
@@ -1523,7 +1523,7 @@ static void R_SortDrawSurfs( drawSurf_t *drawSurfs, int numDrawSurfs ) {
 		dlight_t *dl;
 		// all the lit surfaces are in a single queue
 		// but each light's surfaces are sorted within its subsection
-		for ( i = 0; i < tr.refdef.num_dlights; ++i ) { 
+		for ( i = 0; i < (int)tr.refdef.num_dlights; ++i ) { 
 			dl = &tr.refdef.dlights[ i ];
 			if ( dl->head ) {
 				R_SortLitsurfs( dl );
